extracted highlight behavior to its own class
